Move WKImageRecommendationsView to separate file
