+++ Components/Sources/Components/Components/Suggested Edits/Image Recommendations/WKImageRecommendationsViewController.swift
@@ -13,42 +13,6 @@ fileprivate final class WKImageRecommendationsHostingViewController: WKComponent
 
 }
 
-struct WKImageRecommendationsView: View {
-    
-    @ObservedObject var viewModel: WKImageRecommendationsViewModel
-    @State private var loading: Bool = false
-    
-    var body: some View {
-        Group {
-            if let articleSummary = viewModel.currentRecommendation?.articleSummary {
-                VStack {
-                    Text(articleSummary.displayTitle)
-                    Button(action: {
-                        loading = true
-                        viewModel.next {
-                            loading = false
-                        }
-                    }, label: {
-                        Text("Next")
-                    })
-                }
-            } else {
-                if !loading {
-                    Text("Empty")
-                } else {
-                    ProgressView()
-                }
-            }
-        }
-        .onAppear {
-            loading = true
-            viewModel.fetchImageRecommendations {
-                loading = false
-            }
-        }
-    }
-}
-
 public final class WKImageRecommendationsViewController: WKCanvasViewController {
     
     // MARK: - Properties
